move profile fixtures into conftest
