+++ backend/test/conftest.py
@@ -629,3 +629,192 @@ def echo_server():
     print(f"Echo server terminating", flush=True)
     p.terminate()
     print(f"Echo server terminated", flush=True)
+
+
+PROFILE_NAME = "Test profile"
+PROFILE_DESC = "Profile used for backend tests"
+
+PROFILE_NAME_UPDATED = "Updated test profile"
+PROFILE_DESC_UPDATED = "Updated profile used for backend tests"
+
+PROFILE_2_NAME = "Second test profile"
+PROFILE_2_DESC = "Second profile used to test list endpoint"
+
+def prepare_browser_for_profile_commit(
+    browser_id: str, headers: Dict[str, str], oid: UUID
+) -> None:
+    # Ping to make sure it doesn't expire
+    r = requests.post(
+        f"{API_PREFIX}/orgs/{oid}/profiles/browser/{browser_id}/ping",
+        headers=headers,
+    )
+    assert r.status_code == 200
+    data = r.json()
+    assert data.get("success")
+    assert data.get("origins") or data.get("origins") == []
+
+    # Verify browser seems good
+    r = requests.get(
+        f"{API_PREFIX}/orgs/{oid}/profiles/browser/{browser_id}",
+        headers=headers,
+    )
+    assert r.status_code == 200
+    data = r.json()
+    assert data["url"]
+    assert data["path"]
+    assert data["password"]
+    assert data["auth_bearer"]
+    assert data["scale"]
+    assert data["oid"] == oid
+
+    # Navigate to new URL
+    r = requests.post(
+        f"{API_PREFIX}/orgs/{oid}/profiles/browser/{browser_id}/navigate",
+        headers=headers,
+        json={"url": "https://webrecorder.net/tools"},
+    )
+    assert r.status_code == 200
+    assert r.json()["success"]
+
+    # Ping browser until ready
+    max_attempts = 20
+    attempts = 1
+    while attempts <= max_attempts:
+        try:
+            r = requests.post(
+                f"{API_PREFIX}/orgs/{oid}/profiles/browser/{browser_id}/ping",
+                headers=headers,
+            )
+            data = r.json()
+            if data["success"]:
+                break
+            time.sleep(5)
+        except:
+            pass
+        attempts += 1
+
+
+@pytest.fixture(scope="module")
+def profile_id(admin_auth_headers, default_org_id, profile_browser_id):
+    prepare_browser_for_profile_commit(
+        profile_browser_id, admin_auth_headers, default_org_id
+    )
+
+    # Create profile
+    start_time = time.monotonic()
+    time_limit = 300
+    while True:
+        try:
+            r = requests.post(
+                f"{API_PREFIX}/orgs/{default_org_id}/profiles",
+                headers=admin_auth_headers,
+                json={
+                    "browserid": profile_browser_id,
+                    "name": PROFILE_NAME,
+                    "description": PROFILE_DESC,
+                },
+                timeout=10,
+            )
+            assert r.status_code == 200
+            data = r.json()
+            if data.get("detail") and data.get("detail") == "waiting_for_browser":
+                time.sleep(5)
+                continue
+            if data.get("added"):
+                assert data["storageQuotaReached"] in (True, False)
+                return data["id"]
+        except:
+            if time.monotonic() - start_time > time_limit:
+                raise
+            time.sleep(5)
+
+
+@pytest.fixture(scope="module")
+def profile_config_id(admin_auth_headers, default_org_id, profile_id):
+    r = requests.get(
+        f"{API_PREFIX}/orgs/{default_org_id}/profiles/{profile_id}",
+        headers=admin_auth_headers,
+    )
+    assert r.status_code == 200
+    data = r.json()
+    assert data["id"] == profile_id
+    assert data["name"] == PROFILE_NAME
+    assert data["description"] == PROFILE_DESC
+    assert data["userid"]
+    assert data["oid"] == default_org_id
+    assert data.get("origins") or data.get("origins") == []
+    assert data["createdBy"]
+    assert data["createdByName"] == "admin"
+    assert data["modifiedBy"]
+    assert data["modifiedByName"] == "admin"
+    assert not data["baseid"]
+
+    created = data["created"]
+    assert created
+    assert created.endswith("Z")
+
+    modified = data["modified"]
+    assert modified
+    assert modified.endswith("Z")
+
+    resource = data["resource"]
+    assert resource
+    assert resource["filename"]
+    assert resource["hash"]
+    assert resource["size"]
+    assert resource["storage"]
+    assert resource["storage"]["name"]
+    assert resource.get("replicas") or resource.get("replicas") == []
+
+    # Use profile in a workflow
+    r = requests.post(
+        f"{API_PREFIX}/orgs/{default_org_id}/crawlconfigs/",
+        headers=admin_auth_headers,
+        json={
+            "runNow": False,
+            "name": "Profile Test Crawl",
+            "description": "Crawl using browser profile",
+            "config": {
+                "seeds": [{"url": "https://webrecorder.net/"}],
+                "exclude": "community",
+            },
+            "profileid": profile_id,
+        },
+    )
+    data = r.json()
+    return data["id"]
+
+
+@pytest.fixture(scope="module")
+def profile_2_id(admin_auth_headers, default_org_id, profile_browser_2_id):
+    prepare_browser_for_profile_commit(
+        profile_browser_2_id, admin_auth_headers, default_org_id
+    )
+
+    # Create profile
+    start_time = time.monotonic()
+    time_limit = 300
+    while True:
+        try:
+            r = requests.post(
+                f"{API_PREFIX}/orgs/{default_org_id}/profiles",
+                headers=admin_auth_headers,
+                json={
+                    "browserid": profile_browser_2_id,
+                    "name": PROFILE_2_NAME,
+                    "description": PROFILE_2_DESC,
+                },
+                timeout=10,
+            )
+            assert r.status_code == 200
+            data = r.json()
+            if data.get("detail") and data.get("detail") == "waiting_for_browser":
+                time.sleep(5)
+            if data.get("added"):
+                assert data["storageQuotaReached"] in (True, False)
+
+                return data["id"]
+        except:
+            if time.monotonic() - start_time > time_limit:
+                raise
+            time.sleep(5)
